refactored all MMIF related fn to app.py
